+++ elfi/examples/gauss.py
@@ -58,17 +58,17 @@ def get_model(n_obs=50, true_params=None, seed_obs=None):
 
     """
     if true_params is None:
-        true_params = [10, 2]
+        true_params = [1, 1]
 
     y_obs = Gauss(*true_params, n_obs=n_obs, random_state=np.random.RandomState(seed_obs))
     sim_fn = partial(Gauss, n_obs=n_obs)
 
     m = elfi.ElfiModel()
-    elfi.Prior('uniform', -10, 50, model=m, name='mu')
+    elfi.Prior('uniform', -2, 4, model=m, name='mu')
     elfi.Prior('truncnorm', 0.01, 5, model=m, name='sigma')
-    elfi.Simulator(sim_fn, m['mu'], m['sigma'], observed=y_obs, name='Gauss')
-    elfi.Summary(ss_mean, m['Gauss'], name='S1')
-    elfi.Summary(ss_var, m['Gauss'], name='S2')
-    elfi.Distance('euclidean', m['S1'], m['S2'], name='d')
+    elfi.Simulator(sim_fn, m['mu'], m['sigma'], observed=y_obs, name='gm')
+    elfi.Summary(ss_mean, m['gm'], name='ss_mean')
+    elfi.Summary(ss_var, m['gm'], name='ss_var')
+    elfi.Distance('euclidean', m['ss_mean'], m['ss_var'], name='d')
 
     return m